move CreateSkillSuggestionSheet to its own file
